+++ app/src/main/java/ru/vlasov/carmanager/network/json/main/request/CarEngineRequest.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 class CarEngineRequest(
-        @SerialName("cylinder_volume_in_litres") val cylinderVolumeInLitres : Int,
-        @SerialName("engine_power") val enginePower : Int,
+        @SerialName("cylinder_volume_in_litres") val cylinderVolumeInLitres : Double,
+        @SerialName("engine_power") val enginePower : Double,
         @SerialName("type") val type : String)
